+++ src/3d/cameraRef.ts
@@ -4,6 +4,7 @@ export type CameraRefValue = {
   position: { x: number; z: number }
   rotationY: number
   isFlyingTo: boolean
+  isInVR: boolean
 }
 
 export type CameraRef = MutableRefObject<CameraRefValue>
@@ -13,5 +14,6 @@ export function createCameraRefDefault(): CameraRefValue {
     position: { x: 0, z: 5 },
     rotationY: 0,
     isFlyingTo: false,
+    isInVR: false,
   }
 }